requirements.txt
Flask==3.1.2
mysql-connector-python==9.4.0
python-dotenv==1.1.1
requests==2.32.5
openai==1.30.0
reportlab==4.1.3
python-docx==0.8.12
openpyxl==3.2.2
pandas==2.1.1
click==8.2.1
colorama==0.4.6
greenlet==3.2.4
idna==3.10
charset-normalizer==3.4.3
certifi==2025.8.3
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
itsdangerous==2.2.0
typing_extensions==4.15.0
urllib3==2.5.0
